switch back to single rollup config
