+++ draft-js-buttons/package.json
@@ -29,7 +29,7 @@
   "scripts": {
     "clean": "../node_modules/.bin/rimraf lib",
     "build": "yarn clean && yarn build:js && yarn build:dts",
-    "build:js": "../node_modules/.bin/rollup --config ../rollup.config.ts.js",
+    "build:js": "../node_modules/.bin/rollup --config ../rollup.config.js",
     "build:dts": "tsc -d",
     "prepublish": "yarn build"
   },
