+++ rollup.config.js
@@ -1,9 +1,13 @@
 import path from 'path';
 import nodeResolve from 'rollup-plugin-node-resolve';
 import babel from 'rollup-plugin-babel';
+import { existsSync } from 'fs';
 
-const input = './src/index.js';
-const external = id => !id.startsWith('.') && !path.isAbsolute(id);
+const input = existsSync('./src/index.ts')
+  ? './src/index.ts'
+  : './src/index.tsx';
+const external = (id) => !id.startsWith('.') && !path.isAbsolute(id);
+const extensions = ['.ts', '.js', '.tsx', '.jsx'];
 
 export default [
   {
@@ -14,7 +18,10 @@ export default [
       exports: 'named',
     },
     external,
-    plugins: [nodeResolve(), babel({ rootMode: 'upward' })],
+    plugins: [
+      nodeResolve({ extensions }),
+      babel({ rootMode: 'upward', extensions }),
+    ],
   },
   {
     input,
@@ -24,8 +31,9 @@ export default [
     },
     external,
     plugins: [
-      nodeResolve(),
+      nodeResolve({ extensions }),
       babel({
+        extensions,
         rootMode: 'upward',
         plugins: [
           [
